+++ src/Sylius/Bundle/PromotionBundle/Resources/config/services.xml
@@ -13,11 +13,7 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <imports>
-        <import resource="services/eligibility_checkers.xml" />
-        <import resource="services/criteria.xml" />
-        <import resource="services/forms.xml" />
-        <import resource="services/command.xml" />
-        <import resource="services/validators.xml" />
+        <import resource="services/*.xml" />
     </imports>
 
     <services>
