+++ src/Sylius/Bundle/PromotionBundle/Resources/config/services/console_command.xml
@@ -2,10 +2,12 @@
     <services>
         <defaults public="true" />
 
-        <service id="Sylius\Bundle\PromotionBundle\Command\GenerateCouponsCommand">
+        <service id="Sylius\Bundle\PromotionBundle\Console\Command\GenerateCouponsCommand">
             <argument type="service" id="sylius.repository.promotion" />
             <argument type="service" id="sylius.promotion_coupon_generator" />
             <tag name="console.command" command="sylius:promotion:generate-coupons"/>
         </service>
+
+        <service id="Sylius\Bundle\PromotionBundle\Command\GenerateCouponsCommand" alias="Sylius\Bundle\PromotionBundle\Console\Command\GenerateCouponsCommand" />
     </services>
 </container>
